Move audit app to new project structure
