+++ django/apps/accounts/ldap.py
@@ -1,23 +1,8 @@
-from collections import defaultdict
-from typing import Callable, Any
-
 from django.contrib.auth.models import Group
 
 from apps.accounts import logger
 from apps.accounts.models import User
-
-
-class keydefaultdict(defaultdict):
-    default_factory: Callable[[str], Any]
-
-    # TODO: remove after feature/audit-log merge
-    def __missing__(self, key):
-        if self.default_factory is None:
-            raise KeyError(key)
-
-        ret = self[key] = self.default_factory(key)
-        return ret
-
+from apps.utils.utils import keydefaultdict
 
 _group_cache = keydefaultdict(lambda n: Group.objects.get_or_create(name=n)[0])
