+++ django/apps/utils/utils.py
@@ -1,6 +1,16 @@
+from collections import defaultdict
 from datetime import datetime
 
 from django.utils.formats import get_format
+
+
+class keydefaultdict(defaultdict):
+    def __missing__(self, key):
+        if self.default_factory is None:
+            raise KeyError(key)
+        else:
+            ret = self[key] = self.default_factory(key)
+            return ret
 
 
 def parse_date(value):
