--- web/apps/utils/utils.py
@@ -1,10 +1,0 @@
-from collections import defaultdict
-
-
-class keydefaultdict(defaultdict):
-    def __missing__(self, key):
-        if self.default_factory is None:
-            raise KeyError(key)
-        else:
-            ret = self[key] = self.default_factory(key)
-            return ret
